Bug fixes
I was in a bit of a hurry in the first pass. I missed a massive bug.

When batch deleting, you must translate the pointers (with lowerbound) before any deletions take place.
This suggests a better alogrithm than sort/remove and that is erase/remove.

I did the first pass from memory, and I'm sure there is a way to remove/insert in one pass, using sort, but I can't remember it.

My fix not only removes the bug, but is faster, so win/win. The fact I can't remember the cool trick with sort is annoying, but not required.